refactor : 기능별 폴더 생성
