+++ src/components/roadmap/editDueDateModal.jsx
@@ -1,5 +1,5 @@
 import { useEffect, useState } from "react";
-import "../css/roadmap/editDueDateModal.css";
+import "../../css/roadmap/editDueDateModal.css";
 
 /**
  * 진행 중인 미션의 완료 예정일을 수정하는 모달 컴포넌트
