+++ src/components/roadmap/resultMoveBtn.jsx
@@ -1,5 +1,5 @@
 import { useNavigate } from "react-router-dom";
-import "../css/roadmap/resultBtn.css";
+import "../../css/roadmap/resultBtn.css";
 
 // 로드맵 완료 후 결과 페이지로 이동하는 버튼 컴포넌트
 function ResultBtn() {
