+++ src/components/worldcup/CategorySelector.jsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from "react";
 import { useNavigate } from "react-router-dom";
-import { getCategories, selectJobsByCategory } from "../api/worldcupApi";
-import '../css/CategorySelector.css';
+import { getCategories, selectJobsByCategory } from "../../api/worldcup/worldcupApi";
+import '../../css/worldcup/CategorySelector.css';
 
 export default function CategorySelector() {
     const [categories, setCategories] = useState([]);
