+++ src/css/worldcup/CategorySelector.css
